Refactor user authentication flow and improve error handling in login and registration
