+++ public/css/login.css
@@ -25,14 +25,14 @@ body {
   margin-bottom: 1.5rem;
 }
 
-.input-group {
+  .input-group {
   display: flex;
   align-items: center;
   border: 1px solid #444;
   border-radius: 5px;
   padding: 0.5rem;
   background-color: #1e1e1e;
-  margin-bottom: 1rem;
+  margin-bottom: 0.5rem; /* تقليل المسافة السفلية */
 }
 
 .input-group i {
@@ -46,6 +46,19 @@ body {
   background: transparent;
   color: white;
   width: 100%;
+}
+
+/* تنسيق لرسالة الخطأ */
+#email_err,
+#pass_err
+{
+  display: block;
+  color: red;
+  font-size: 14px;
+  margin-top: 5px; /* مسافة قليلة بين الحقل ورسالة الخطأ */
+  word-wrap: break-word; /* للسماح بلف النص بشكل جيد */
+  white-space: normal; /* التأكد من أن النص سيتكسر إذا كان طويلًا */
+  margin-bottom: 0; /* إزالة المسافة السفلية غير المطلوبة */
 }
 
 .checkbox-container {
