+++ routes/birds.js
@@ -1,6 +1,7 @@
 const express = require('express')
 const router = express.Router()
  const usercontroller = require("../Controllers/usercontroller")
+ const authcontrller = require("../Controllers/authcontrller")
  var {requierdAuth} =require("../middleware/middleware")
  var {checkuser} = require ("../middleware/middleware")
 // get الخاصه هذه اكواد المشروع الحقيقي
@@ -8,13 +9,6 @@ const router = express.Router()
  router.get("*" ,checkuser)
 
 router.get("/" , checkuser, usercontroller.user_welcom_get)
-
-router.get("/singout", usercontroller.user_signout_get)
-
-
-router.get("/login", usercontroller.user_login_get)
-
-router.get("/register", usercontroller.user_register_get)
 
 router.get("/home", requierdAuth,  usercontroller.user_index_get )
 
@@ -33,9 +27,18 @@ router.put("/edit/:id", usercontroller.user_edit_put)
 
 
  // هنا بيكون الخاص بحق انشاءحساب جديد 
- router.post("/register", usercontroller.user_register_post)
+ router.post("/register", authcontrller.user_register_post)
 
- router.post("/login", usercontroller.user_login_post)
+ router.post("/login", authcontrller.user_login_post)
+
+ 
+router.get("/singout", authcontrller.user_signout_get)
+
+
+router.get("/login", authcontrller.user_login_get)
+
+router.get("/register", authcontrller.user_register_get)
+
 
 // app.post("/", (req, res) => {
 //   console.log(req.body);
